# Modules/Purchase/routes/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Purchase Management Module API",
    "description": "Comprehensive Purchase Management API for Branch Managers - Handles order creation, tracking, receiving goods, and return management",
    "version": "1.0.0",
    "contact": {
      "name": "Assab API Support"
    }
  },
  "servers": [
    {
      "url": "/api/v1/purchase",
      "description": "Purchase Module API Base URL"
    }
  ],
...
  "paths": {
...
    "/history/{id}": {
      "get": {
        "tags": ["Purchase History"],
        "summary": "Get order details",
        "description": "Retrieve detailed information about a specific order",
        "operationId": "getOrderDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
    "/history/{id}/timeline": {
      "get": {
        "tags": ["Purchase History"],
        "summary": "Get order timeline",
        "description": "Retrieve complete audit trail for an order",
        "operationId": "getOrderTimeline",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order timeline retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
...
    "/pending": {
      "get": {
        "tags": ["Pending Orders"],
        "summary": "Get pending orders",
        "description": "Retrieve list of pending orders with filters",
        "operationId": "getPendingOrders",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Order type filter",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["direct_supplier", "via_purchasing_officer", "internal_transfer", "transfer_received", "multiple_sources"]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Status filter",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["draft", "pending", "pending_confirmation", "pending_approval", "partial_confirmation", "confirmed", "preparing", "on_the_way", "delayed"]
            }
          },
          {
            "name": "date_range",
            "in": "query",
            "description": "Date range filter",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["last_24h", "last_7d", "last_30d", "custom"]
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 15
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPurchaseOrders"
                }
              }
            }
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
...
    "/pending/{id}/reject": {
      "post": {
        "tags": ["Pending Orders"],
        "summary": "Reject order",
        "description": "Reject a pending order",
        "operationId": "rejectOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["reason"],
                "properties": {
                  "reason": {
                    "type": "string",
                    "maxLength": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order rejected successfully"
          }
        },
        "security": [{"bearerAuth": []}]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "data": {
            "type": "object"
          },
          "meta": {
            "type": "object",
            "properties": {
              "pagination": {
                "type": "object"
              },
              "filters_applied": {
                "type": "object"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "type": "object"
          }
        }
      },
...
      "PaginatedPurchaseOrders": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PurchaseOrder"
                }
              },
              "links": {
                "type": "object",
                "properties": {
                  "first": {
                    "type": "string"
                  },
                  "last": {
                    "type": "string"
                  },
                  "prev": {
                    "type": "string",
                    "nullable": true
                  },
                  "next": {
                    "type": "string",
                    "nullable": true
                  }
                }
              },
              "meta": {
                "type": "object",
                "properties": {
                  "current_page": {
                    "type": "integer"
                  },
                  "from": {
                    "type": "integer",
                    "nullable": true
                  },
                  "last_page": {
                    "type": "integer"
                  },
                  "per_page": {
                    "type": "integer"
                  },
                  "to": {
                    "type": "integer",
                    "nullable": true
                  },
                  "total": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        ]
      },
      "PurchaseOrder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "order_number": {
            "type": "string",
            "example": "DS-20251207-ABCD"
          },
          "order_type": {
            "type": "string",
            "enum": ["direct_supplier", "via_purchasing_officer", "internal_transfer", "multiple_sources", "transfer_received"]
          },
          "order_type_label": {
            "type": "string",
            "example": "Direct Supplier Order"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "pending", "pending_confirmation", "pending_approval", "partial_confirmation", "confirmed", "preparing", "on_the_way", "delivered", "closed", "canceled", "rejected", "delayed"]
          },
          "status_label": {
            "type": "string",
            "example": "Pending"
          },
          "status_color": {
            "type": "string",
            "example": "#F59E0B"
          },
          "priority": {
            "type": "string",
            "enum": ["high", "normal"]
          },
          "quality_level": {
            "type": "string",
            "enum": ["economy", "standard", "premium"],
            "nullable": true
          },
          "total_items": {
            "type": "integer"
          },
          "received_items": {
            "type": "integer"
          },
          "subtotal": {
            "type": "number",
            "format": "float"
          },
          "tax_amount": {
            "type": "number",
            "format": "float"
          },
          "total_amount": {
            "type": "number",
            "format": "float"
          },
          "preferred_delivery_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "latest_delivery_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}